requirements.txt
APScheduler==3.7.0
blinker==1.4
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
dnspython==1.16.0
eventlet==0.25.2
Flask==1.1.2
Flask-Admin==1.5.6
Flask-APScheduler==1.11.0
Flask-Cors==3.0.8
Flask-Excel==0.0.7
Flask-HTTPAuth==4.2.0
Flask-Mail==0.9.1
Flask-SocketIO==4.3.0
Flask-SQLAlchemy==2.4.1
greenlet==0.4.15
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
lml==0.1.0
MarkupSafe==1.1.1
monotonic==1.5
numpy==1.19.4
pandas==1.1.5
pyexcel==0.6.6
pyexcel-io==0.6.4
pyexcel-webio==0.1.4
python-dateutil==2.8.1
python-engineio==3.12.1
python-socketio==4.5.1
pytz==2020.4
requests==2.23.0
six==1.14.0
SQLAlchemy==1.3.17
SQLAlchemy-Utils==0.36.8
texttable==1.6.3
tzlocal==2.1
urllib3==1.25.9
Werkzeug==1.0.1
WTForms==2.3.1
